Snippets: redesign for JavaScript, C and C++
Create a "cfamily" snippets file that contains what looks the same in C, C++ and
JavaScript code. Then create a snippet for C that will contain what is not
exactly the same with JavaScript (e.g. a for in C is usually "int i" instead of
"var i"), then the ones specific for C++ in its own file.

There is still the need to find a way of splitting Qt only snippets to its own
file.