Split gitconfig in two files
Now the usual ~/.gitconfig has the usual name and email, but after that just
includes a ~/.gitconfig.local that has the real meat. This way my coworkers can
clone my repo, and tell git to set their name and email just fine, without
changing a file under version control and not committing it, because gitconfig
is copied and only .gitconfig.local is symlinked.