Updates to submodules (unhappy state)
This commit updates lots of submodules, but not all are fully up to date, since
I could not test everything. This is what happens when you want to live on the
edge, and then go one month on vacation. :)

Added indentLine plugin, so disable indent_guides.

Also, enable markdown folding (support added in this new version, yay).