Bump fugitive version
This is "just" an update bringing more than 800 commits, spanning more
than 2 and a half years... And not that much has been changed in a
breaking way, which is quite impressive. The next commits seem to remove
the support of some mappings that I'm still getting used to avoid, so
I'll probably be here for a bit more.