Update qml plugin, and disable its folding
The folding of the plugin is pretty 'meh', as it just folds on
functions, but the main deal to me is folding elements, so just disable
it and keep using folding on indentation levels.

While I have 'indent' as default for foldmethod, set it in a ftplugin
for QML files.