Big plugin update, adding mini.nvim and other stuff
It was due to start using more "modern" plugins seriously. Mini seems
convenient in that a lot of the functionality is the usual one that
seems pretty "core" (though I'm not necessarily activating that just
yet). It also is gonna save me from the scare of having so many things
on the runtimepath/packpath/whatever.

This also removes VimFiler, which for some reason started failing after
the update to 0.10 (and it's not worth the time updating it, as VimFiler
is deprecated by its author). I'll try MiniFiles for now.

Also, adds the "helpful" plugin, which should be pretty much a default.
And I've managed to add it "on demand", by using a ftplugin file in Lua
and some tricks. Let's see if they hold the test of time.