Remove Plenary and Telescope
I wasn't even using it, given that I've been so slow moving away from
the old Vim plugins, and this was in the `opt` subdirectory, so it
wasn't even loaded. Whenever I get rid of Unite, we'll see what I end up
using.

But preemptively removing it given that I don't want to use anything
that reminds me of bigots. See: https://archive.is/6S4tA